fixed post method

+++ utopia-flights-api/src/main/java/com/start/flights/api/repository/FlightRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.start.flights.api.model.Flight;
 
 @Repository
-public interface FlightRepository extends JpaRepository<Flight, Long> {
+public interface FlightRepository extends JpaRepository<Flight, Integer> {
 	
 }
